moves low-level stuff into libogg.jl
also cleans up some of the object creation/destruction
