+++ src/Ogg.jl
@@ -1,10 +1,13 @@
 __precompile__()
 module Ogg
-using FileIO
-import Base: show, convert
-export OggDecoder, OggEncoder, OggPage
-export serial, eos, bos
-export eachpage
+
+export OggDecoder
+
+# using FileIO
+# import Base: show, convert
+# export OggDecoder, OggEncoder, OggPage
+# export serial, eos, bos
+# export eachpage
 
 const depfile = joinpath(@__DIR__, "..", "deps", "deps.jl")
 if isfile(depfile)
@@ -13,8 +16,8 @@ else
     error("libogg not properly installed. Please run Pkg.build(\"Ogg\")")
 end
 
-include("types.jl")
+include("libogg.jl")
 include("decoder.jl")
-include("encoder.jl")
+# include("encoder.jl")
 
 end # module
